refactor: 버킷(Bucket) 비즈니스 로직 개선하기
1. void 메서드를 지양하기 위해 boolean 을 반환하는 checkImageIsEmpty 메서드와 checkImageCount 메서드 추가함.
2. Record 추가 내역 반영함.
3. 외부에서 설정한 이미지 정책 상수값 주입받는 구문 추가함.
